refactor: entity 계층 변경 (infrastructure -> domain)
