+++ src/main/java/com/sparta/hhplusw02cleanarchitecture/domain/entity/LectureItemEntity.java
@@ -1,13 +1,12 @@
-package com.sparta.hhplusw02cleanarchitecture.infrastructure.entity;
+package com.sparta.hhplusw02cleanarchitecture.domain.entity;
 
-import com.sparta.hhplusw02cleanarchitecture.domain.common.TimeBaseEntity;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.Table;
-import java.time.LocalDateTime;
+import java.time.LocalDate;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -15,30 +14,26 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 /**
- * 특강 신청 내역 엔티티
+ * 특강 목록 엔티티
  */
 @Entity
-@Table(name = "LECTURE_HISTORY")
+@Table(name = "LECTURE_ITEM")
 @Getter
 @Builder
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
-public class LectureHistoryEntity extends TimeBaseEntity {
+public class LectureItemEntity {
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
-  @Column(name = "history_id")
+  @Column(name = "item_id")
   private Long id;
-
-  @Column(nullable = false)
-  private Long user_id;
 
   @Column(nullable = false)
   private Long lecture_id;
 
-  @Column(nullable = false)
-  private Long item_id;
+  private LocalDate date;
 
   @Column(nullable = false)
-  private LocalDateTime applied_at;
+  private Integer capacity;
 
 }
