+++ src/main/java/com/sparta/hhplusw02cleanarchitecture/infrastructure/repository/lectureHistory/LectureHistoryRepository.java
@@ -1,7 +1,7 @@
 package com.sparta.hhplusw02cleanarchitecture.infrastructure.repository.lectureHistory;
 
 
-import com.sparta.hhplusw02cleanarchitecture.infrastructure.entity.LectureHistoryEntity;
+import com.sparta.hhplusw02cleanarchitecture.domain.entity.LectureHistoryEntity;
 import org.springframework.stereotype.Repository;
 
 /**
