+++ src/main/java/com/sparta/hhplusw02cleanarchitecture/infrastructure/repository/lectureInventory/LectureInventoryJpaRepository.java
@@ -1,6 +1,6 @@
 package com.sparta.hhplusw02cleanarchitecture.infrastructure.repository.lectureInventory;
 
-import com.sparta.hhplusw02cleanarchitecture.infrastructure.entity.LectureInventoryEntity;
+import com.sparta.hhplusw02cleanarchitecture.domain.entity.LectureInventoryEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 /**
