+++ src/main/java/com/sparta/hhplusw02cleanarchitecture/infrastructure/repository/lectureInventory/LectureInventoryRepositoryJpaImpl.java
@@ -1,6 +1,6 @@
 package com.sparta.hhplusw02cleanarchitecture.infrastructure.repository.lectureInventory;
 
-import com.sparta.hhplusw02cleanarchitecture.infrastructure.entity.LectureInventoryEntity;
+import com.sparta.hhplusw02cleanarchitecture.domain.entity.LectureInventoryEntity;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
