+++ src/test/java/com/sparta/hhplusw02cleanarchitecture/domain/lecture/usecase/ApplyLectureServiceTest.java
@@ -4,8 +4,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.when;
 
-import com.sparta.hhplusw02cleanarchitecture.infrastructure.entity.LectureHistoryEntity;
-import com.sparta.hhplusw02cleanarchitecture.infrastructure.entity.LectureInventoryEntity;
+import com.sparta.hhplusw02cleanarchitecture.domain.entity.LectureHistoryEntity;
+import com.sparta.hhplusw02cleanarchitecture.domain.entity.LectureInventoryEntity;
 import com.sparta.hhplusw02cleanarchitecture.infrastructure.repository.lecture.LectureQueryRepository;
 import com.sparta.hhplusw02cleanarchitecture.infrastructure.repository.lectureHistory.LectureHistoryRepositoryJpaImpl;
 import com.sparta.hhplusw02cleanarchitecture.infrastructure.repository.lectureInventory.LectureInventoryRepositoryJpaImpl;
